api/pyproject.toml
[project]
name = "api"
version = "0.2.0"
description = "API to transform documents into universal ontology"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.111.0",
    "fastapi-utils>=0.6.0",
    "uvicorn>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "pydantic<2.0.0",
    "jsonpath-ng>=1.6.1",
    "requests>=2.32.2",
    "sentence-transformers>=3.0.0",
    "sentencepiece>=0.2.0",
    "torch>=2.3.0",
    "protobuf>=5.27.0",
    "elasticsearch>=8.6.2",
    "beautifulsoup4>=4.12.3",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "bcrypt>=4.3.0",
    "python-dotenv==1.0.1",
    "huggingface-hub[hf-xet]>=0.30.2",
    "python-multipart>=0.0.20",
    "python-slugify>=8.0.4",
    "pytest-mock>=3.14.0",
]

[dependency-groups]
dev = [
    "black>=24.4.2",
    "pytest>=8.2.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1"
]

[tool.black]
line-length = 128

[tool.pytest.ini_options]
pythonpath = ["."]
